Move variable classes from workload to new module
Moves variable classes out of workload module now that they are part of
shared language.
